Separate out library API and makle constants file
