+++ packages/webapp/src/store/reducer.js
@@ -18,6 +18,7 @@ import baseReducer from '../containers/reducer';
 import { combineForms } from 'react-redux-form';
 import { PURGE } from 'redux-persist';
 import { api } from './api/apiSlice';
+import { libraryApi } from './api/libraryApiSlice';
 import insightReducer from '../containers/Insights/reducer';
 import financeReducer from '../containers/Finances/reducer';
 import certifierSurveyReducer from '../containers/OrganicCertifierSurvey/slice';
@@ -236,6 +237,7 @@ const tempStateReducer = combineReducers({
 // combine all reducers here and pass it to application
 const appReducer = combineReducers({
   [api.reducerPath]: api.reducer,
+  [libraryApi.reducerPath]: libraryApi.reducer,
   profileForms: combineForms(
     {
       addInfo: addUserInfo,
